02_math_agent_w_fastmcp/requirements.txt
google-adk >= 1.0.0 
google-genai == 1.16.1
asyncio == 3.4.3
asyncstdlib == 3.13.1

# agent engine
google-cloud-aiplatform[adk,agent_engines]

# mcp
fastmcp == 2.5.1
mcp == 1.9.1

# testing/evaluation
pytest == 8.3.5
dotenv == 0.9.9
scikit-learn == 1.6.1
tabulate == 0.9.0
